Add better overseer-test-prompt history.
This patch adds a separate history list for `overseer-test-prompt`.
First time user runs this, they have no choice but to enter the prompt
text.  Next time however, the previous input is made the default, so
they can simply RET to confirm to re-run the previous input.

Using `M-n` and `M-p` the user can browse the prompt history.  Selecting
any item from the history makes it the default for the next run.